+++ slurm/logs/train/what am i training.txt
@@ -1,0 +1,27 @@
+@ Monday Sept 29, 2025
+UME protein small
+shared with Taylor
+s3://prescient-lobster/ume/runs/2025-09-26T13-51-56/epoch=1-step=98000-val_loss=0.4485.ckpt
+https://genentech.wandb.io/zadorozk/lobster/runs/88a7pgtu
+
+---
+UME protein medium
+119073 try with small warmup stage and 5e-4
+https://genentech.wandb.io/zadorozk/lobster/runs/upnvtr0p
+
+-- 
+UME smiles small ZINC
+1 gpu for now
+121066 
+https://genentech.wandb.io/zadorozk/lobster/runs/979k47vg
+
+UME smiles small MolPile
+121128
+https://genentech.wandb.io/zadorozk/lobster/runs/ytvukehh
+
+
+-- 
+UME interaction
+stopped
+somehow cosine with restarts
+https://genentech.wandb.io/zadorozk/lobster/runs/y47nwlyf?nw=nwuserzadorozk
