+++ src/lobster/datasets/s3_datasets/__init__.py
@@ -5,7 +5,7 @@ from .calm import Calm
 from .m320m import M320M
 from .peptide_atlas import PeptideAtlas
 from .zinc import ZINC
-from .random_interactions import RandomInteractions
+from .molpile import MolPILE
 
 __all__ = [
     "AMPLIFY",
@@ -15,5 +15,5 @@ __all__ = [
     "ZINC",
     "UMEStreamingDataset",
     "Atomica",
-    "RandomInteractions",
+    "MolPILE",
 ]
