+++ src/lobster/hydra_config/experiment/ume-2/small_molecule.yaml
@@ -8,13 +8,13 @@ defaults:
 
 compile: false
 
-rdkit_descriptors:
-  - MolWt
-  - NumHAcceptors
-  - NumHDonors
-  - TPSA
-  - SlogP_VSA6
-  - BertzCT
+# rdkit_descriptors:
+#   - MolWt
+#   - NumHAcceptors
+#   - NumHDonors
+#   - TPSA
+#   - SlogP_VSA6
+#   - BertzCT
 
 data:
   _target_: lobster.data.UMELightningDataModule
@@ -26,50 +26,54 @@ data:
   batch_size: 64
   pin_memory: true
   num_workers: 8
-  dataset_kwargs:
-    ZINC:
-      use_shared_tokenizer: false
-      extra_transform_fns:
-        rdkit:
-          _target_: lobster.transforms.SmilesToRDKitDescriptorsTransform
-          normalize: true
-          invert: true
-          descriptor_list: ${rdkit_descriptors}
+  use_shared_tokenizer: false
+  # dataset_kwargs:
+  #   ZINC:
+  #     use_shared_tokenizer: false
+  #     extra_transform_fns:
+  #       rdkit:
+  #         _target_: lobster.transforms.SmilesToRDKitDescriptorsTransform
+  #         normalize: true
+  #         invert: true
+  #         descriptor_list: ${rdkit_descriptors}
 
 model:
   _target_: lobster.model.ume2.UMESequenceEncoderLightningModule 
-  auxiliary_tasks:
-    - _target_: lobster.model.ume2.AuxiliaryTask
-      name: rdkit
-      task_type: regression
-      output_dim: 6
-      hidden_size: null
-      loss_weight: 0.05
-      pooling: mean
-      num_layers: 2
-      dropout: 0.1
+  # auxiliary_tasks:
+  #   - _target_: lobster.model.ume2.AuxiliaryTask
+  #     name: rdkit
+  #     task_type: regression
+  #     output_dim: 6
+  #     hidden_size: null
+  #     loss_weight: 0.05
+  #     pooling: cls
+  #     num_layers: 2
+  #     dropout: 0.1
   lr: 1e-5
-  scheduler: constant_with_warmup
+  scheduler: cosine
   scheduler_kwargs:
-    num_warmup_steps: 0
+    num_warmup_steps: 1_000
+    num_training_steps: 100_000
   weight_decay: 0.01
   mask_token_id: 4
   pad_token_id: 1
   special_token_ids: [0, 1, 2, 3, 4, 5]
   mask_probability: 0.3
-  ckpt_path: null
+  ckpt_path: #s3://prescient-lobster/ume/runs/${paths.timestamp}/last.ckpt
   encoder_kwargs:
     max_length: ${data.max_length}
-    model_ckpt: s3://prescient-lobster/ume/runs/2025-09-03T15-31-34/epoch=2-step=89000-val_loss=0.8960.ckpt
     cache_dir: /data2/ume/checkpoints
     model_size: small
     vocab_size: 1226
 
 logger:
-  name: UME-2-${model.encoder_kwargs.model_size}_${data.datasets}_sweep_${paths.timestamp}
+  name: UME2-${model.encoder_kwargs.model_size}_${data.datasets}_smiles_no_aux_${paths.timestamp}
   project: lobster
   entity: ${oc.env:LOBSTER_USER}
   save_dir: ${oc.env:LOBSTER_RUNS_DIR}
+  # id: fvxv9lff
+  # resume: must
+  # allow_val_change: true
 
 paths:
   timestamp: ${now:%Y-%m-%d}T${now:%H-%M-%S} 
@@ -81,7 +85,7 @@ trainer:
   max_epochs: -1
   gradient_clip_val: 0.5
   max_time: null
-  max_steps: -1
+  max_steps: ${model.scheduler_kwargs.num_training_steps}
   val_check_interval: 500
   limit_val_batches: 50_000
   precision: 16-mixed
